add docker script as alternative to fig
@sckott I've added this script because fig doesn't always seem to work
how I would expect. This just calls `docker run` for each of the
containers manually instead.  The issues with fig will probably be
resolved when docker's own orchestration tool (compose) is out of alpha;
then we can just do `docker up`. Anyway, this script is there if you
need it.